requirements.txt
MouseInfo~=0.1.3
Pillow~=7.2.0
PyAutoGUI~=0.9.50
PyGetWIndow~=0.0.8
PyMsgBox~=1.0.9
PyRect~=0.1.4
PyScreeze~=0.1.26
PyTweening~=1.0.3
numpy~=1.19.2
opencv-python~=4.4.0.42
pynput~=1.7.1
pyperclip~=1.8.0
python-imageseach-drov0~=1.0.6
python3-xlib~=0.15
six~=1.15.0